카카오로그인 수정, 필요없는 파일 삭제

+++ springTeamProject/src/main/java/sol/one/VO/BoardlistVO.java
@@ -9,7 +9,7 @@ public class BoardlistVO {
 	
 	private int product_id;
 	
-	private int user_id;
+	private long user_id;
 	
 	private Date board_write_time;
