+++ springTeamProject/src/main/java/sol/one/controller/ProductController.java
@@ -47,7 +47,7 @@ public class ProductController {
 	private TradeLogService tradelogservice;
 
 	@GetMapping("/get/getMySellPdstatus")
-	public Map<String,Object> getMySellPdstatus(int product_id,int sell_user_id) {
+	public Map<String,Object> getMySellPdstatus(int product_id,long sell_user_id) {
 		Map<String,Object> map = new HashMap<>();
 		T_tradeVO vo = tradelogservice.getsellmypdstatus(product_id,sell_user_id);
 		if(vo == null) {
@@ -63,7 +63,7 @@ public class ProductController {
 	}
 	
 	@GetMapping("/get/getMyBuyPdstatus")
-	public Map<String,Object> getMyBuyPdstatus(int product_id,int sell_user_id,int buyer_user_id) {
+	public Map<String,Object> getMyBuyPdstatus(int product_id,long sell_user_id,long buyer_user_id) {
 		Map<String,Object> map = new HashMap<>();
 		T_tradeVO vo = tradelogservice.getbuymypdstatus(product_id,sell_user_id,buyer_user_id);
 		if(vo == null) {
@@ -79,7 +79,7 @@ public class ProductController {
 	}
 	
 	@GetMapping("/getMyLike")
-	public Map<String,Object> getMyLike(int product_id,int user_id) {
+	public Map<String,Object> getMyLike(int product_id,long user_id) {
 		Map<String,Object> map = new HashMap<>();
 		LikeVO vo = tradelogservice.getLike(product_id, user_id);
 		if(vo == null) {
