+++ springTeamProject/src/main/java/sol/one/controller/WriteController.java
@@ -54,7 +54,7 @@ public class WriteController {
 	//등록버튼을 눌렀을때 
 	@PostMapping("/product_add")
 	public String add(String pd_name,String pd_img,String pd_desc,String company_name,String pd_price,
-			int category_id,int location_id,int pd_discount,String pd_date,int user_id,String pd_price0) {
+			int category_id,int location_id,int pd_discount,String pd_date,long user_id,String pd_price0) {
 		ProductVO vo = new ProductVO();
 		vo.setPd_desc(pd_desc);
 		vo.setPd_img(pd_img);
